requirements.txt
google-generativeai
gradio
python-dotenv
Pillow